+++ src/io/react-native/react-native-dummy.js
@@ -1,0 +1,7 @@
+// @flow
+
+import type { EdgeRawIo } from '../../edge-core-index.js'
+
+export function makeReactNativeIo (): Promise<EdgeRawIo> {
+  throw new Error('This function only works on React Native')
+}